Updates to sys_command.sh
Added wake capabilities, a custom command and logging. Wake.pl does the
same thing, but I liked having everything in one basket.

The custom command allows for commands that are not in the script to be
activated. Ex: sys_command custom all. If “ls” the script will ssh into
the location argument and run the command. Handy for on the fly changes.

The log file catches who runs the command and what was typed into the
custom command since the history doesn’t capture it.